More tidy
