+++ tembo-operator/src/extensions/database_queries.rs
@@ -13,6 +13,10 @@ lazy_static! {
     static ref VALID_INPUT: Regex = Regex::new(r"^[a-zA-Z]([a-zA-Z0-9]*[-_]?)*[a-zA-Z0-9]+$").unwrap();
 }
 
+pub fn check_input(input: &str) -> bool {
+    VALID_INPUT.is_match(input)
+}
+
 pub const LIST_DATABASES_QUERY: &str = r#"SELECT datname FROM pg_database WHERE datistemplate = false;"#;
 
 pub const LIST_EXTENSIONS_QUERY: &str = r#"select
@@ -278,6 +282,97 @@ pub async fn toggle_extension(
     Ok(())
 }
 
-pub fn check_input(input: &str) -> bool {
-    VALID_INPUT.is_match(input)
+
+#[cfg(test)]
+mod tests {
+    use crate::extensions::database_queries::{check_input, parse_databases, parse_extensions};
+
+    #[test]
+    fn test_parse_databases() {
+        let three_db = " datname
+        ----------
+         postgres
+         cat
+         dog
+        (3 rows)
+
+         ";
+
+        let rows = parse_databases(three_db);
+        println!("{:?}", rows);
+        assert_eq!(rows.len(), 3);
+        assert_eq!(rows[0], "postgres");
+        assert_eq!(rows[1], "cat");
+        assert_eq!(rows[2], "dog");
+
+        let one_db = " datname
+        ----------
+         postgres
+        (1 row)
+
+         ";
+
+        let rows = parse_databases(one_db);
+        println!("{:?}", rows);
+        assert_eq!(rows.len(), 1);
+        assert_eq!(rows[0], "postgres");
+    }
+
+    #[test]
+    fn test_parse_extensions() {
+        let ext_psql = "        name        | version | enabled |   schema   |                              description
+        --------------------+---------+---------+------------+------------------------------------------------------------------------
+         adminpack          | 2.1     | f       | public     | administrative functions for PostgreSQL
+         amcheck            | 1.3     | f       | public     | functions for verifying relation integrity
+         autoinc            | 1.0     | f       | public     | functions for autoincrementing fields
+         bloom              | 1.0     | f       | public     | bloom access method - signature file based index
+         btree_gin          | 1.3     | f       | public     | support for indexing common datatypes in GIN
+         btree_gist         | 1.7     | f       | public     | support for indexing common datatypes in GiST
+         citext             | 1.6     | f       | public     | data type for case-insensitive character strings
+         cube               | 1.5     | f       | public     | data type for multidimensional cubes
+         dblink             | 1.2     | f       | public     | connect to other PostgreSQL databases from within a database
+         (9 rows)";
+
+        let ext = parse_extensions(ext_psql);
+        assert_eq!(ext.len(), 9);
+        assert_eq!(ext[0].name, "adminpack");
+        assert!(!ext[0].enabled);
+        assert_eq!(ext[0].version, "2.1".to_owned());
+        assert_eq!(ext[0].schema, "public".to_owned());
+        assert_eq!(
+            ext[0].description,
+            "administrative functions for PostgreSQL".to_owned()
+        );
+
+        assert_eq!(ext[8].name, "dblink");
+        assert!(!ext[8].enabled);
+        assert_eq!(ext[8].version, "1.2".to_owned());
+        assert_eq!(ext[8].schema, "public".to_owned());
+        assert_eq!(
+            ext[8].description,
+            "connect to other PostgreSQL databases from within a database".to_owned()
+        );
+    }
+
+    #[test]
+    fn test_check_input() {
+        let invalids = ["extension--", "data;", "invalid^#$$characters", ";invalid", ""];
+        for i in invalids.iter() {
+            assert!(!check_input(i), "input {} should be invalid", i);
+        }
+
+        let valids = [
+            "extension_a",
+            "schema_abc",
+            "extension",
+            "NewExtension",
+            "NewExtension123",
+            "postgis_tiger_geocoder-3",
+            "address_standardizer-3",
+            "xml2",
+        ];
+        for i in valids.iter() {
+            assert!(check_input(i), "input {} should be valid", i);
+        }
+    }
 }
