+++ tembo-operator/src/extensions/mod.rs
@@ -1,109 +1,25 @@
 pub mod database_queries;
+pub mod install;
 pub mod kubernetes_queries;
-pub mod trunk;
+pub mod toggle;
 pub mod types;
-pub mod extensions;
 
 use crate::{
-    apis::coredb_types::CoreDB, Context, controller::patch_cdb_status_merge, get_current_coredb_resource,
+    apis::coredb_types::CoreDB,
+    extensions::types::{ExtensionStatus, TrunkInstallStatus},
+    Context,
 };
-use kube::{api::Api, runtime::controller::Action};
-use serde_json::json;
+use kube::{runtime::controller::Action, Api};
 use std::sync::Arc;
-use tracing::error;
-use types::{Extension, ExtensionInstallLocationStatus, ExtensionStatus, TrunkInstallStatus};
 
-#[cfg(test)]
-mod tests {
-
-    use crate::extensions::database_queries::{check_input, parse_databases, parse_extensions};
-
-    #[test]
-    fn test_parse_databases() {
-        let three_db = " datname  
-        ----------
-         postgres
-         cat
-         dog
-        (3 rows)
-        
-         ";
-
-        let rows = parse_databases(three_db);
-        println!("{:?}", rows);
-        assert_eq!(rows.len(), 3);
-        assert_eq!(rows[0], "postgres");
-        assert_eq!(rows[1], "cat");
-        assert_eq!(rows[2], "dog");
-
-        let one_db = " datname  
-        ----------
-         postgres
-        (1 row)
-        
-         ";
-
-        let rows = parse_databases(one_db);
-        println!("{:?}", rows);
-        assert_eq!(rows.len(), 1);
-        assert_eq!(rows[0], "postgres");
-    }
-
-    #[test]
-    fn test_parse_extensions() {
-        let ext_psql = "        name        | version | enabled |   schema   |                              description                               
-        --------------------+---------+---------+------------+------------------------------------------------------------------------
-         adminpack          | 2.1     | f       | public     | administrative functions for PostgreSQL
-         amcheck            | 1.3     | f       | public     | functions for verifying relation integrity
-         autoinc            | 1.0     | f       | public     | functions for autoincrementing fields
-         bloom              | 1.0     | f       | public     | bloom access method - signature file based index
-         btree_gin          | 1.3     | f       | public     | support for indexing common datatypes in GIN
-         btree_gist         | 1.7     | f       | public     | support for indexing common datatypes in GiST
-         citext             | 1.6     | f       | public     | data type for case-insensitive character strings
-         cube               | 1.5     | f       | public     | data type for multidimensional cubes
-         dblink             | 1.2     | f       | public     | connect to other PostgreSQL databases from within a database
-         (9 rows)";
-
-        let ext = parse_extensions(ext_psql);
-        assert_eq!(ext.len(), 9);
-        assert_eq!(ext[0].name, "adminpack");
-        assert!(!ext[0].enabled);
-        assert_eq!(ext[0].version, "2.1".to_owned());
-        assert_eq!(ext[0].schema, "public".to_owned());
-        assert_eq!(
-            ext[0].description,
-            "administrative functions for PostgreSQL".to_owned()
-        );
-
-        assert_eq!(ext[8].name, "dblink");
-        assert!(!ext[8].enabled);
-        assert_eq!(ext[8].version, "1.2".to_owned());
-        assert_eq!(ext[8].schema, "public".to_owned());
-        assert_eq!(
-            ext[8].description,
-            "connect to other PostgreSQL databases from within a database".to_owned()
-        );
-    }
-
-    #[test]
-    fn test_check_input() {
-        let invalids = ["extension--", "data;", "invalid^#$$characters", ";invalid", ""];
-        for i in invalids.iter() {
-            assert!(!check_input(i), "input {} should be invalid", i);
-        }
-
-        let valids = [
-            "extension_a",
-            "schema_abc",
-            "extension",
-            "NewExtension",
-            "NewExtension123",
-            "postgis_tiger_geocoder-3",
-            "address_standardizer-3",
-            "xml2",
-        ];
-        for i in valids.iter() {
-            assert!(check_input(i), "input {} should be valid", i);
-        }
-    }
+/// reconcile extensions between the spec and the database
+pub async fn reconcile_extensions(
+    coredb: &CoreDB,
+    ctx: Arc<Context>,
+    _cdb_api: &Api<CoreDB>,
+    _name: &str,
+) -> Result<(Vec<TrunkInstallStatus>, Vec<ExtensionStatus>), Action> {
+    let trunk_installs = install::reconcile_trunk_installs(coredb, ctx.clone()).await?;
+    let extension_statuses = toggle::reconcile_extension_toggle_state(coredb, ctx.clone()).await?;
+    Ok((trunk_installs, extension_statuses))
 }
